* Commit de respaldo

+++ spring/udemy-springsecurity6/springsecsection/src/main/java/com/eazybytes/repository/CustomerRepository.java
@@ -1,14 +1,15 @@
 package com.eazybytes.repository;
 
-import com.eazybytes.model.Customer;
+import java.util.List;
+
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
+import com.eazybytes.model.Customer;
 
 @Repository
-public interface CustomerRepository extends CrudRepository<Customer,Long> {
+public interface CustomerRepository extends CrudRepository<Customer, Integer> {
 
-    List<Customer> findByEmail(String email);
-    
+	List<Customer> findByEmail(String email);
+
 }
